+++ app/src/main/java/com/stopsmoke/kekkek/presentation/MainActivity.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
-import com.stopsmoke.kekkek.HomeFragment
+import com.stopsmoke.kekkek.presentation.home.HomeFragment
 import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.databinding.ActivityMainBinding
 import com.stopsmoke.kekkek.presentation.community.CommunityFragment
