+++ app/src/main/java/com/stopsmoke/kekkek/presentation/home/test/TestFragment.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek
+package com.stopsmoke.kekkek.presentation.home.test
 
 import android.os.Bundle
 import android.view.LayoutInflater
