+++ app/src/main/java/com/stopsmoke/kekkek/presentation/home/test/TestOnBoardingFragment.kt
@@ -1,13 +1,10 @@
-package com.stopsmoke.kekkek
+package com.stopsmoke.kekkek.presentation.home.test
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.commit
-import androidx.viewpager2.widget.ViewPager2
-import com.stopsmoke.kekkek.databinding.FragmentTestBinding
 import com.stopsmoke.kekkek.databinding.FragmentTestOnboardingBinding
 
 class TestOnBoardingFragment(
