+++ app/src/main/java/com/stopsmoke/kekkek/presentation/home/test/ViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek
+package com.stopsmoke.kekkek.presentation.home.test
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
