+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:padding="20dp"
     android:background="@drawable/bg_home"
-    tools:context=".HomeFragment">
+    tools:context=".presentation.home.HomeFragment">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
