+++ app/src/main/res/layout/fragment_question.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#1E1E1E"
-    tools:context=".QuestionFragment">
+    tools:context=".presentation.home.test.QuestionFragment">
 
 
     <TextView
