+++ app/src/main/res/layout/fragment_test.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TestFragment">
+    tools:context=".presentation.home.test.TestFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewpager_test"
