+++ app/src/main/res/layout/fragment_test_onboarding.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#A2AAD1"
-    tools:context=".TestOnBoardingFragment">
+    tools:context=".presentation.home.test.TestOnBoardingFragment">
 
 
     <ImageView
